backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
alembic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
